+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 
 ## [1.7.5] - 2023-05-28
+### Added
+- SuccessMessageMixin to TeamNewEditBaseView
 ### Changed
 - RaceDetailView to RaceTahminView
 - Refactor NewTahminView